antenna-code/station-tools/requirements.txt
argcomplete==3.5.3
bleak==0.22.3
certifi==2025.1.31
charset-normalizer==3.4.1
dbus-fast==2.33.0
dotmap==1.3.30
idna==3.10
meshtastic==2.5.12
packaging==24.2
print-color==0.4.6
protobuf==5.29.3
pynmea2==1.19.0
pynmeagps==1.0.44
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
pytap2==2.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
RouterOS-api==0.18.1
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
